+++ src/test/java/CalculatorTest.java
@@ -12,32 +12,32 @@ class CalculatorTest {
 
     @Test
     void plus() throws OperationNotSupportedException {
-        Integer plus = SimpleCalculator.calculate(PLUS);
+        Double plus = SimpleCalculator.calculate(PLUS);
         assertEquals(5, plus);
     }
 
     @Test
     void minus() throws OperationNotSupportedException {
-        Integer minus = SimpleCalculator.calculate(MINUS);
+        Double minus = SimpleCalculator.calculate(MINUS);
         assertEquals(4, minus);
     }
 
     @Test
     void multi() throws OperationNotSupportedException {
-        Integer multi = SimpleCalculator.calculate(MULTI);
+        Double multi = SimpleCalculator.calculate(MULTI);
         assertEquals(14, multi);
     }
 
     @Test
     void deli() throws OperationNotSupportedException {
-        Integer deli = SimpleCalculator.calculate(DELI);
+        Double deli = SimpleCalculator.calculate(DELI);
         assertEquals(3, deli);
     }
 
     @Test
     void multiCalcPlus() throws OperationNotSupportedException {
         String multiEx = "3 + 3 + 3 + 3";
-        Integer result = SimpleCalculator.calculate(multiEx);
+        Double result = SimpleCalculator.calculate(multiEx);
 
         assertEquals(12, result);
     }
@@ -45,7 +45,7 @@ class CalculatorTest {
     @Test
     void multiCalcMinus() throws OperationNotSupportedException {
         String multiEx = "3 - 3 - 3 - 3";
-        Integer result = SimpleCalculator.calculate(multiEx);
+        Double result = SimpleCalculator.calculate(multiEx);
 
         assertEquals(-6, result);
     }
@@ -53,7 +53,7 @@ class CalculatorTest {
     @Test
     void multiCalc() throws OperationNotSupportedException {
         String multiEx = "3 * 3 / 3 + 3 - 3";
-        Integer result = SimpleCalculator.calculate(multiEx);
+        Double result = SimpleCalculator.calculate(multiEx);
 
         assertEquals(3, result);
     }
@@ -61,7 +61,7 @@ class CalculatorTest {
     @Test
     void multiCalcHard() throws OperationNotSupportedException {
         String ex = "4 * 2 / 4 + 21";
-        Integer res = SimpleCalculator.calculate(ex);
+        Double res = SimpleCalculator.calculate(ex);
 
         assertEquals(23, res);
     }
@@ -69,7 +69,7 @@ class CalculatorTest {
     @Test
     void calculate1() throws OperationNotSupportedException {
         String ex = "2 + 2 * 2";
-        Integer res = SimpleCalculator.calculate(ex);
+        Double res = SimpleCalculator.calculate(ex);
 
         assertEquals(6, res);
     }
@@ -77,7 +77,7 @@ class CalculatorTest {
     @Test
     void calculate2() throws OperationNotSupportedException {
         String ex = "2 + 2 * 2 / 4 + 6 - 1 + 2 * 3";
-        Integer res = SimpleCalculator.calculate(ex);
+        Double res = SimpleCalculator.calculate(ex);
 
         assertEquals(14, res);
     }
@@ -85,7 +85,7 @@ class CalculatorTest {
     @Test
     void calculateHard1() throws OperationNotSupportedException {
         String ex = "(2 + 2) * 2 / 4 + ((6 - 1 + 2) * 3)";
-        Integer res = SimpleCalculator.calculate(ex);
+        Double res = SimpleCalculator.calculate(ex);
 
         assertEquals(23, res);
     }
@@ -93,7 +93,7 @@ class CalculatorTest {
     @Test
     void calculateHard2() throws OperationNotSupportedException {
         String ex = "2 + ((2 * 2) / 4) + 6 - 1 + 2 * 3";
-        Integer res = SimpleCalculator.calculate(ex);
+        Double res = SimpleCalculator.calculate(ex);
 
         assertEquals(14, res);
     }
